Remove the package installation helpers + fixes
Also:
Add helpers for Ubuntu 18.04
Update Amazon Linux to install 17.12
